Tests fixed
